переработка чтения конфигурации
