+++ ToDoList.PL/appsettings.json
@@ -1,5 +1,7 @@
 {
   "configurationDal": {
-    "Type":  1
+    "Type": 1,
+    "FileName": "Base.json",
+    "FilePath":  ""
   }
 }
