+++ ToDoList.sln
@@ -27,8 +27,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ToDoList.Exceptions", "ToDo
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ToDoList.IOC", "ToDoList.IOC\ToDoList.IOC.csproj", "{1F646931-FFAF-48B8-AFCB-E79BBA6083ED}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ToDoList.Settings", "ToDoList.Settings\ToDoList.Settings.csproj", "{0CA3D9B4-FDD4-4308-94BB-6936FAAD91A5}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ToDoList.PL", "ToDoList.PL\ToDoList.PL.csproj", "{A82EC361-FD81-4AFF-B060-E5555968927E}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ToDoList.PL.Views", "ToDoList.PL.Views\ToDoList.PL.Views.csproj", "{90EB84E5-BF37-4576-9F2E-4E9E574CDFF2}"
@@ -75,10 +73,6 @@ Global
 		{1F646931-FFAF-48B8-AFCB-E79BBA6083ED}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1F646931-FFAF-48B8-AFCB-E79BBA6083ED}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1F646931-FFAF-48B8-AFCB-E79BBA6083ED}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0CA3D9B4-FDD4-4308-94BB-6936FAAD91A5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0CA3D9B4-FDD4-4308-94BB-6936FAAD91A5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0CA3D9B4-FDD4-4308-94BB-6936FAAD91A5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0CA3D9B4-FDD4-4308-94BB-6936FAAD91A5}.Release|Any CPU.Build.0 = Release|Any CPU
 		{A82EC361-FD81-4AFF-B060-E5555968927E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A82EC361-FD81-4AFF-B060-E5555968927E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A82EC361-FD81-4AFF-B060-E5555968927E}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -108,7 +102,6 @@ Global
 		{BA3CA880-E41B-43D7-AB87-DA9EC96ADC9E} = {C88A44ED-3877-40E8-BE87-D03709BA08EC}
 		{E4EEEE9D-7D60-4579-B340-360D3908A510} = {21F851ED-A98E-44B2-98D8-57CAC6D3314E}
 		{1F646931-FFAF-48B8-AFCB-E79BBA6083ED} = {21F851ED-A98E-44B2-98D8-57CAC6D3314E}
-		{0CA3D9B4-FDD4-4308-94BB-6936FAAD91A5} = {21F851ED-A98E-44B2-98D8-57CAC6D3314E}
 		{A82EC361-FD81-4AFF-B060-E5555968927E} = {8B084800-216A-4731-A840-E6565E5AD5E8}
 		{90EB84E5-BF37-4576-9F2E-4E9E574CDFF2} = {8B084800-216A-4731-A840-E6565E5AD5E8}
 		{C8BB30C8-2FDA-4E6D-A2C3-9B318C49B174} = {8B084800-216A-4731-A840-E6565E5AD5E8}
